requirements.txt
# Core Django Framework
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Authentication & Security
django-allauth==0.57.0
django-otp==1.2.2
django-guardian==2.4.0
django-crypto-fields==0.1.3
django-admin-honeypot==1.1.0

# Database & Storage
# psycopg2-binary==2.9.9  # For PostgreSQL - install separately if needed
redis==5.0.1
django-redis==5.4.0

# Async & Real-time
celery==5.3.4
channels==4.2.2
channels-redis==4.2.1

# Search & Analytics
elasticsearch==8.11.1
django-elasticsearch-dsl==7.3

# File Storage
django-storages==1.14.2
boto3==1.34.0

# API & Documentation
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.27.0
graphene-django==3.2.0

# Utilities
python-decouple==3.8
Pillow==10.1.0
python-slugify==8.0.1
django-extensions==3.2.3

# AI/ML Integration
transformers==4.36.2
torch==2.1.2
scikit-learn==1.3.2
numpy==1.24.4
pandas==2.1.4

# Monitoring & Logging
sentry-sdk==1.39.2
django-debug-toolbar==4.2.0

# Testing
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2

# Production
gunicorn==21.2.0
whitenoise==6.6.0
